# Conflicts:
#	inventory-system/src/pages/products/ProductForm.vue
#	inventory-system/src/store/modules/category.js

+++ inventory-system/src/store/modules/category.js
@@ -1,4 +1,5 @@
 import Vue from 'vue'
+
 const state = {
   categories: [],
   categoryId: '',
@@ -16,47 +17,50 @@ const getters = {
 const mutations = {
   setCategories: (state, categories) => {
     state.categories = categories;
-	},
-	// removeCategory: (state, categoryId) =>{
-	// 	state.categories.delete(categoryId)
-	// }
+  },
+
+
+  // removeCategory: (state, categoryId) =>{
+  // 	state.categories.delete(categoryId)
+  // }
   // setCategoryId: (state, categories) =>{
-	//   state.categoryId = categoryId;
+  //   state.categoryId = categoryId;
   // }
 };
 
 
-
 const actions = {
 
-	initCategory: ({commit}) => {
-		return Vue.axios.get('/api/category/list').then((response)=>{
-			commit('setCategories', response.data);
-			return (response.data)
-	})
-	},
+  initCategory: ({commit}) => {
+    return Vue.axios.get('/api/category/list').then((response) => {
+      commit('setCategories', response.data);
+      return (response.data)
+    })
+  },
 
-	setCategories: ({commit},categories) => {
-		commit('setCategories', categories );
-	},
+  setCategories: ({commit}, categories) => {
+    commit('setCategories', categories);
+  },
 
-	DELETE_CATEGORY: ({commit}, categoryId)=>{
-		return new Promise((resolve, reject)=> {
-			Vue.axios.delete('/api/category/delete?id='+ categoryId).then((response) =>{
-				Vue.axios.get('/api/category/list').then((response)=>{
-					console.log(response.data)
-					commit('setCategories', response.data)
-					resolve(response.data)
-				})
-			})
-		})
-	}
-};
+  DELETE_CATEGORY: ({commit}, categoryId) => {
+    if (confirm('Delete?')) {
+      return new Promise((resolve, reject) => {
+        Vue.axios.delete('/api/category/delete?id=' + categoryId).then((response) => {
+          Vue.axios.get('/api/category/list').then((response) => {
+            console.log(response.data)
+            commit('setCategories', response.data)
+            resolve(response.data)
+          })
+        })
+      })
+    }
+  }
+}
 
 
-export default{
-	state,
-	mutations,
-	actions,
-	getters
+export default {
+  state,
+  mutations,
+  actions,
+  getters
 }
